+++ rasa_core/channels/telegram.py
@@ -8,6 +8,7 @@ from telegram import (
 from rasa_core import constants, utils
 from rasa_core.channels import InputChannel
 from rasa_core.channels.channel import UserMessage, OutputChannel
+from rasa_core.constants import INTENT_MESSAGE_PREFIX, USER_INTENT_RESTART
 
 logger = logging.getLogger(__name__)
 
@@ -149,7 +150,7 @@ class TelegramInput(InputChannel):
                         return response.text("success")
                 sender_id = msg.chat.id
                 try:
-                    if text in {'_restart', constants.USER_INTENT_RESTART}:
+                    if text == (INTENT_MESSAGE_PREFIX + USER_INTENT_RESTART):
                         await on_new_message(UserMessage(
                             text, out_channel, sender_id,
                             input_channel=self.name()))