+++ rasa_core/processor.py
@@ -1,33 +1,29 @@
-import time
-
 import json
 import logging
-import numpy as np
 from types import LambdaType
-from typing import Optional, List, Dict, Any, Tuple
-from typing import Text
+from typing import Any, Dict, List, Optional, Text, Tuple
+
+import numpy as np
+import time
 
 from rasa_core import jobs
 from rasa_core.actions import Action
 from rasa_core.actions.action import (
     ACTION_LISTEN_NAME,
-    ACTION_RESTART_NAME,
     ActionExecutionRejection)
-from rasa_core.channels import CollectingOutputChannel
-from rasa_core.channels import UserMessage
+from rasa_core.channels import CollectingOutputChannel, UserMessage
+from rasa_core.constants import (
+    ACTION_NAME_SENDER_ID_CONNECTOR_STR,
+    USER_INTENT_RESTART)
 from rasa_core.dispatcher import Dispatcher
 from rasa_core.domain import Domain
-from rasa_core.events import ReminderScheduled, Event
-from rasa_core.events import SlotSet
 from rasa_core.events import (
-    UserUttered,
-    ActionExecuted,
-    BotUttered,
-    ActionExecutionRejected)
+    ActionExecuted, ActionExecutionRejected,
+    BotUttered, Event, ReminderCancelled, ReminderScheduled, SlotSet,
+    UserUttered)
 from rasa_core.interpreter import (
-    NaturalLanguageInterpreter,
-    INTENT_MESSAGE_PREFIX)
-from rasa_core.interpreter import RegexInterpreter
+    INTENT_MESSAGE_PREFIX,
+    NaturalLanguageInterpreter, RegexInterpreter)
 from rasa_core.nlg import NaturalLanguageGenerator
 from rasa_core.policies.ensemble import PolicyEnsemble
 from rasa_core.tracker_store import TrackerStore
@@ -270,10 +266,11 @@ class MessageProcessor(object):
         logger.debug("Logged UserUtterance - "
                      "tracker now has {} events".format(len(tracker.events)))
 
-    def _should_handle_message(self, tracker):
+    @staticmethod
+    def _should_handle_message(tracker):
         return (not tracker.is_paused() or
                 tracker.latest_message.intent.get("name") ==
-                self.domain.restart_intent)
+                USER_INTENT_RESTART)
 
     async def _predict_and_execute_next_action(self, message, tracker):
         # keep taking actions decided by the policy until it chooses to 'listen'
@@ -321,22 +318,40 @@ class MessageProcessor(object):
         return not is_listen_action
 
     async def _schedule_reminders(self, events: List[Event],
+                                  tracker: DialogueStateTracker,
                                   dispatcher: Dispatcher) -> None:
         """Uses the scheduler to time a job to trigger the passed reminder.
 
         Reminders with the same `id` property will overwrite one another
         (i.e. only one of them will eventually run)."""
 
-        if events is not None:
-            for e in events:
-                if isinstance(e, ReminderScheduled):
-                    (await jobs.scheduler()).add_job(
-                        self.handle_reminder, "date",
-                        run_date=e.trigger_date_time,
-                        args=[e, dispatcher],
-                        id=e.name,
-                        replace_existing=True,
-                        name=str(e.action_name))
+        for e in events:
+            if isinstance(e, ReminderScheduled):
+                (await jobs.scheduler()).add_job(
+                    self.handle_reminder, "date",
+                    run_date=e.trigger_date_time,
+                    args=[e, dispatcher],
+                    id=e.name,
+                    replace_existing=True,
+                    name=(str(e.action_name) +
+                          ACTION_NAME_SENDER_ID_CONNECTOR_STR +
+                          tracker.sender_id))
+
+    @staticmethod
+    async def _cancel_reminders(events: List[Event],
+                                tracker: DialogueStateTracker) -> None:
+        """Cancel reminders by action_name"""
+
+        # All Reminders with the same action name will be cancelled
+        for e in events:
+            if isinstance(e, ReminderCancelled):
+                name_to_check = (str(e.action_name) +
+                                 ACTION_NAME_SENDER_ID_CONNECTOR_STR +
+                                 tracker.sender_id)
+                scheduler = await jobs.scheduler()
+                for j in scheduler.get_jobs():
+                    if j.name == name_to_check:
+                        scheduler.remove_job(j.id)
 
     async def _run_action(self, action, tracker, dispatcher, policy=None,
                           confidence=None):
@@ -360,7 +375,9 @@ class MessageProcessor(object):
         self._log_action_on_tracker(tracker, action.name(), events, policy,
                                     confidence)
         self.log_bot_utterances_on_tracker(tracker, dispatcher)
-        await self._schedule_reminders(events, dispatcher)
+
+        await self._schedule_reminders(events, tracker, dispatcher)
+        await self._cancel_reminders(events, tracker)
 
         return self.should_predict_another_action(action.name(), events)
 
@@ -449,10 +466,12 @@ class MessageProcessor(object):
         else:
             return None, None
 
-    def _get_next_action_probabilities(
-        self,
-        tracker: DialogueStateTracker
-    ) -> Tuple[Optional[List[float]], Optional[Text]]:
+    def _get_next_action_probabilities(self,
+                                       tracker: DialogueStateTracker
+                                       ) -> Tuple[Optional[List[float]],
+                                                  Optional[Text]]:
+        """Collect predictions from ensemble and return action and predictions.
+        """
 
         followup_action = tracker.followup_action
         if followup_action:
@@ -466,9 +485,5 @@ class MessageProcessor(object):
                     "Instead of running that, we will ignore the action "
                     "and predict the next action.".format(followup_action))
 
-        latest_intent = tracker.latest_message.intent.get("name")
-        if latest_intent == self.domain.restart_intent:
-            return self._prob_array_for_action(ACTION_RESTART_NAME)
-        else:
-            return self.policy_ensemble.probabilities_using_best_policy(
-                tracker, self.domain)
+        return self.policy_ensemble.probabilities_using_best_policy(
+            tracker, self.domain)